Ignore certain commands in history
Always ignore 1 letter commands.  Ignore 2 letter commands if
given with no arguments (I am not sure when we might want to
keep 2 letter commands with args, but am playing cautious.)
Ignore commands which clearly have no historical value.
(For now, that is only pwd.  We can add to the list.)