Use C-p to toggle arrows
Trying to use Up and Down fails miserably when the mouse kicks
in and sends lots of events.  It does not seem to happen with
left and right.  This is a bit of a kludge, but allows the operator
to manually toggle the keys (so they will become active in the
embedded session running on the remote.)