Remove empty history file
I do not know what is happening.  On macos, I continually get in a state
where bash commands are not getting written to the empty file.  I do
not know how to list the open filedescriptors of the shell in this POS
os, so use this sledge hammer to try to fix the issue.