Do not remove gsutil, gcloud, bq expansions
Often, I end up commenting these out on machines where the expansions
are not defined.  Minimalism would say we should delete these, practicality
says it is really annoying when these expansions are in place so I will
keep them commented to make it easier to recover when I hit the pause.