Suppress errors from pwd
Still getting lots of:

"shell-init: error retrieving current directory: getcwd: cannot access parent directories: No such file or directory"

but this eliminates the error messages I seem to be responsible for.  The system
startup scripts on macos are a total disaster, but I have not yet
tracked down the source of the shell-init error.  I hope it is not
directly from bash, but am not going to waste any more time
tracking it down.  A shell should be usable even when the underlying
directory is deleted.  This error message is stupid.

Grr:
http://git.savannah.gnu.org/cgit/bash.git/tree/builtins/common.c#n579

That error message comes straight from bash with no way to suppress it.