Abort if not in a git dir
git-diff spwews a usage message instead of an error message when
the branch is empty if we are not in a git repo.  (Note that this
is merely a rant against git.  The correct behavior of any tool
is to report a terse error message, not to spew screenfulls of
usage statement.)