Put tab and - on the extra keys view
We can use the volume-up T shortcut to get tab, but it is annoying.
hyphen is surprisingly not available easily on the hackers keyboard,
but I have stopped using that keyboard anyway so that is not relevant.
But hyphen is not availalbe with a volume-up shortcut, so it needs
to be accessible.  I never use Alt, so get rid of it.