Turn off pull.rebase
I am often seeing a git pull just be a git fetch, and
I have to explicitly reset --hard @{u}, which seems
silly.  Since I am more likely to fetch than pull and
reset manually anyway, there does not seem to be much
point having this config set (it is easy enough to
reset if there is ever a merge that I do not want).  Mostly,
I am turning this off to see if the pull continues
to degrade to a mere fetch.