Suppress multiple warnings re: TMUX in new shell
I should track down exactly why the debug trap is executing multiple times when a new
shell opens, but for now this should be good.  It seems that anytime a file is sourced
in the startup sequence, the debug trap is run and the warning is written multiple
times; this is a bit of a hack to suppress the multiple outputs.