Use awk vice grep to parse blacklist
Sigh.  2023 and still having to think about whitespace crap in names
PATH, etc. fundamentally relies on names being sane (ie, not containing
whitespace) but I keep bumping into vendors who do stupid crap like
making the directory in which they put executables have names with
horizontal whitespace, so you have to think about this.  But this is my
own personal dotfiles, so I can impose policy.  And policy is that
names do not contain stupid characters.  So we will not try to make
the code robust.  Grrr, this is annoying.