Explicitly disable formatoptons=o in perl
o in formatoptions is the stupidest feature in existence.  It is already
disabled multiple times in .vimrc, but it still is set when a perl file
is the first opened and the BufEnter autocmd has not yet been run.  Adding
this in ftplugin is a total kludge, and I need to track down the right
way to do this.  Probably the right thing to do is patch vim and
get rid of this stupid option for ever.