Always write history header
Initially, the test -s was there to avoid writing the
header if a shell starts with the same pid as an extant
file, but it seems that is not the behavior we want.  Avoiding
the header would just confuse things if a new shell starts, as
we would not be able to recognize a discontinuity in the
working directory location.