Delete all aliases
I am currently setting up on a RHEL machine where I do not
have root.  The stupid aliases in the system startup files
(in particular, the alias for grep) causes my started files
to have syntax errors.  (grep() { ... } is a syntax error
when grep is expanded as an alias.)  Just purge the idiocy
and uset all aliases as soon as possible.