Highlight color diffs
When removing dos-style line endings from a file (eg 0x0d0a), it
is convenient to be able to see that change.  Note that git diff
will only show whitespace errors on files that are added, so
if you want to see the change when such line endings are removed,
you need to do 'git diff -R' to see the reverse patch.