Do not emit empty trailer header in l1
This is relying on a particular sed implementation.  On the box I am
currently on, sed does not do the replacement at the beginning of the
line when we use printf to emit nothing with no trailing newline.
I believe the failure to emit a newline leads to unspecified sed
behavior, but I do not care.  Worry about it when/if it becomes
an issue.