Clean setup a bit
It is not clear to me if we should just make $HOME/.config
a symbolic link into dotfiles.  On the one hand, we could
clean up some cruft in the setup script by doing that.
OTOH, I don't know what software may write files into
$HOME/config and I don't want to have to constantly
be adding things to .gitignore.  It is not at all
clear to me if config files intended to be sourced
by .vimrc really belong in ~/.config/vim, and that
will be decided at a future date.  (ie, never)