Only do file type and remote check once
The logic here is terrible.  I am trying to get defaults of
expandtab and ts=4 for all files that are related to my
current employer, so I think we should trigger off of the
remote.  But all of this cruft needs to be cleaned up and
taken out of the file type specific logic.  Currently, I
want to keep all go and yaml files at 8 and 4, so this
logic is doing what what I want, but I need to think about
the right way to structure this.  Probably the remote
should come first and not be overridden, but I do not ses
a clean way to do that without leaking names from the remote.

The problem here is that I am trying to mix information from
git and filetype information to decide which settings to use.
The remote of the current branch seems highly relevant.  vim's
startup logic is a bit convoluted, and I don't really want to
have to understand it, but maybe the right thing to do is to
use *.vim files and invoke a script from one of thoe filetype
syntax files.