Do not emit name/email in submodules
When updating a repo with a lot of submodules, this message
becomes spam.  Perhaps we should continue to emit the message
if we are doing an intentional work in the submodule, but I don't
know a way to easily tell if the hook was invoked because of
a command run in the parent module.