Create multiple windows in the new session
Add the max count so that if NEW_SESSION_PANE_COUNT is invalid we
are sure not to spawn too many.  I suspect most invalid values
would just cause sed to error out early, but do not want to think
about it.  The -e 1d is there because the initial new-session
command creates one window, and this is just decrementing the count.