Reverse revs in bundle name, omit root oid
If the base commit does not have a parent, do not include
it in the bundle name.  This should make it clearer when
a bundle has no dependencies.

Also, reverse the order so the base commit is first.  This
seems more intuitive, since it will match the order given
to rev-list.