Stop stripping diff metadata
I had been trying to match with [^[:space:]], but there does not
seem to be a character class of "all spaces except tab", so that
did not really work.  Since we are now just explicitly matching
the space character instead of the class, we do not need to worry
about ^+++ matching agains "^[+]${tab}*[^[:space:]]" or similar.
Attempting to strip those out feels very fragile.

Signed-off-by: William Pursell <william.r.pursell@gmail.com>