Do not filter git branch output with grep
git branch will output to a pager if output is to a tty, and
I like that behavior.  And filtering through grep like this
is icky.  Need to find a better way to hide branches....or
just clean them up.  One option is to just move .git/refs/attic
to .git/refs/remotes/attic, and they'll be hidden by default.

Signed-off-by: William Pursell <william.r.pursell@gmail.com>