Stop using git-diff-index to check indentation
I want space indentation for python files, but hard tabs
for every other file type.  (Really, I want hard tabs in
python as well, but the industry has gone the wrong way
and it's not a battle worth fighting.)  git-diff-index
just doesn't give the flexibility to change indentation
type per file, and works poorly in detecting bad indentation
anyway.  (With core.whitespace set to indent-with-non-tab,
it should flag any indentation using spaces as an error,
but it allows spaces following tabs, and allows indentation
that is all space.  It only flags an error if there is a space
followed by a tab.)

So we'll just parse it ourselves the slow way.