Use <C-l><C-b> to send C-l
Not really sure if this is a good idea.  I like keeping
hjkl for pane navigation, but we need a method for sending
the prefix.  Perhaps just use hjk, since I do not actually
navigate panes that much.  OTOH, perhaps the reason I do
not navigate panes very much is that the bindings are not
easy.  I do not often need to send C-l to the underlying
app, but I suspect the next time I need to redraw a gdb
tui session I will curse myself for having made this
binding.