Reset termio settings
I do not fully understand what is happening here, but it
seems the first invocation of tput does not get an accurate number.
If we create a split pane in tmux, the first call to "tput cols"
returns the width of the original pane, and the width of the
new pane is only given in the second call.  So we pre-heat
the cache by calling tput and discarding the output.