Use \W to get basename of PWD
This gives us ~ substitution for free, which is necessary now
that my HOME directory has such a horrible name.  Note to
corporate IT everywhere: let the user pick their username
on their box.  It does not need to match their corporate
identity.  The output of `basename "$HOME" | wc -c` should
always be 4.