Fail a hook if the index is modified
I have recently encountered a pre-commit that
makes changes to the worktree and invokes git add.
This seems like a terrible violation
of the principle of least surprise, and I do not ever
want hunks to be committed that I am not even aware
of.  The proper flow is to make the changes and then
alert the author of them and allow them to be inspected.