Read .bash-local first
The idea had been to let .bash-local override anything by
running it last, but in practice it is more common to
use it to set variables that need to be set early (eg
PS1_PREFIX).  There used to be two files: one read
first and one read last, and maybe that is a better
way to go.  In practice, right now we just need this
one to be read early.