Print hostname, not PROJECT in .bash-history
I do not seem to be using PROJECT much anymore for anything, so
printing it in the history file seems pointless.  Since the hostname
will be different if we are in a docker container, there is no need
to print DOCKER.  Having the hostname becomes important as we
intend to start merging history files from different hosts.