Inject +v when virtaul environment is active
To be clear, virutalenv is the stupidest pile of crap ever foisted on computer
users.  We keep VIRTUAL_ENV_DISABLE_PROMPT set so that we don't get the full
name of the base in the prompt, which is way too long and ugly to be acceptable,
but I am finding it necessary to distinguish shells that have the fucking thing
enabled.  I thought argparse was the worst thing that Python had ever done,
but this nonsense is more damaging.  Package management in python is such
a cluster-fuck, and this craptastic hack is no solution at all.  Maybe
it's time to just go back to using perl.

Python should be banned from the universe.