Start with all folds open
The current behavior is insanely annoying.  If you open folds in a buffer,
then swith to a different buffer, when you come back to the first all the
folds are again closed.  That is silly.  I want the foldlevel for each
buffer to persist when changing, but they don't.  I'm not sure if this
is because of some setting I currently have in place, or if it is just
the expected behavior.  As a workaround, let's just change the behavior
so that all folds are open when the buffer is entered.  It would be
nice if closed folds persisted instead, but I don't have time right
now to track down how to get that behavior.