Add default behavior for g with no args
git by default acts as "git --help" when no arguments are given, and that's
just plain annoying.  (Correct behavior is to emit an error message.)
Usually, I find that I am only typing g with no arguments when I
fat-finger and miss either the 'b' or the 's', so it seems reasonable
to do both.