Fix gno
This is messed up.  The expectation is (was?) that the first
2 args of gno would (maybe) be branches.  But if the first
arg is --name-only, we do not treat any arg as a branch
but pass it to git diff.  This now works, but messes with the
convention of passing flags before args, and we circumvent
that by shifting.  But it is messed up.