Stop using history -a
It does not work as expected.  I can explicitly invoke "history -a"
and see it return 0 but not add any commands to HISTFILE.  There
is probably a bash bug here, or some odd interaction with my
debug_trap or PROMPT_COMMAND.  In any case, all of this cruft
is fragile, and I am never looking at any of the data, so
just simplify everything.