Replace all numeric args to git with a branch ref
This is a bit fragile, but looking through my bash history I do not
see any instances of raw numbers.  I expected to find a few "-n 10"
or similar, but do not,  but this seems like an obvious gotcha waiting
to happen so we explicitly look for -n.

I believe this is the first instance in which we
are using non-portable bashisms (the array).  So I guess
we are now fully committed to bash vs. sh.  This
functionality seems incredibly useful, and I think
the trade-off is worthwhile.