Get rid of the array
Shell hackery like this is fragile garbage, but removing
a bashism is worth the fragility.  And maybe this is actually
robust! (I do not like manipulating the args with "set" in the
middle of the loop, but for the version of bash I am currently
running this seems to work.