Enhance warning text
The message "/p/a/t/h does not exist" does not inform the
user that it is intended to be a directory.  Perhaps it
is overkill to write 2 distinct message like this and we
could just say "/p/a/t/h/ does not exist or is not a directory",
but we may as well be precise.