Trim leading commas
This allows a global config of core.default-trailers
to be overridden by a local value set to the empty
string. This is a bit kludgy, and I need to clarify
if indeed a comma is the expected delimiter (ie, I
believe I set this up so that default-trailers can
be a comma-separated list, but I don't actually
remember if that was the expected usage.