Remove space after author name
I really do not understand git formats.  If reset is the empty
string, then keeping the space after "an" causes there to
be no space in the output, while there are 2 spaces when
reset is not empty.  However, removing the space causes there
to be exactly one space in each situation.