Suppress annoying advice
I need to read some docs and figure out how to just suppress
all the stupid advice, but just do this for now.  This
is in response to the behavior of git-add when no arguments
are given, which is to print "Nothing specified, nothing added."
to stderr.  But it does not fail!!  Stupid git.