Check all arguments for -n
The headers are incorrect when --date is set, but this at least will
prevent tilde spew from ldh.  The "ldh" alias sets --date and -n,
but -n is not first, and the current implementation of less leaves
a screen full of tildes when --header is applied, despite the
presence of F in LESS, so we disable headers when -n is
applied.

type: git