Allow empty value for trailer
We make "required" mean that it is present, but not non-empty.
The use case here is for the examples repo when I am
adding an entry to .gitignore.  There is a "lang" trailer,
but making a change to a .gitignore does not really have
a language, so it makes sense to leave the trailer emptry.
Maybe we can have config-required-non-empty or something,
but for now we will just allow an empty value.  As long
as the user is explicitly adding something, it indicates
that it is being considered.

type: git