Check all commit messages
In c94e954e193d73df, we added logic to avoid adding trailers
because the trailer was expected to already be there.  However,
the script is now not merely adding a single trailer, but also
doing a lot of validation, and that validation should occur
for all types of commits.

type: git