Move trailer validation from prepare-commit-msg to commit-msg
prepare-commit-msg is used to prepare the message before
the invocation of GIT_EDITOR, so is the correct place to
insert trailers, but we need to give the user a chance
to add trailers if desired.  Therefore we need to do the
validation after the EDITOR is finished, which is in commit-msg.

type: git