Do not parse t in ci alias
This is no longer necessary, since we are now adding the trailers
from t in prepare-commit-msg.  Indeed, this is undesirable since
this mechanism does not parse t as a comma separated list
and just injects the remainder of the list as the value for
the first trailer which is then duplicated in the prepare-commit-msg
hook.

type: git