Use config to grab ticket
I really do not use tickets anymore, but the code is clearly no
longer even checking the config for the branch.  The intent here
is to first check the extant commit message, then the config,
the the local script to find a ticket.  I should probably just
remove all of this logic.

type: bug