Pass extant ref for idempotency
If the ref already exists and is correct, we should allow that.
Passing an empty string as the third arg requires that this be a
new ref, and there really is no requirement for that.  As long as
we do not stomp on any extant ref that is different, it is fine.

type: git