Explicitly exit vice short-circuit
I suppose this is purely a style question.  In general, explicit
is better than implicit, so I like the explicit exit vs. the short
circuit.  Along the same lines, "|| true" indicates that we have
thought about the fact that git show-ref may fail and we explicitly
do not care.  The reason we do not care is that allowing $extant_ref
to be the empty string is exactly the behavior we want.

type: git