Warn if V is set and not an integer
Unclear to me if we should warn whenver V is set.  It is pretty
common to use for verbosity, but maybe a warning is in order.
Usually, I expect to set it for each command and would be
surprised if it were exported, so maybe emit a warning if
it is exported?

type: rc