Check f,h,t,V consistently
It is tempting to check all single letter variables, but I am
a little worried about the slowness of such a check and for the
moment think it better to be selective and only check those
values that I know my own scripts and functions check.  We
can add things when/if they become annoyances.

type: rc