Move gitcommit.vim syntax file
It is not clear why, but I cannot find the file being loaded,
and since the other files are in ftplugin, that seems
like a good place to put it.  Vim startup is a spaghetti
mess, and I do not want to invest time understanding it.
The file gets loaded from ftplugin, so I will put it
there.

type: git