Stop resetting title after commands
I think it is more useful to leave the last command in the
title, rather than reverting to the working directory.
However, for cd, it seems reasonable to continue
setting the title to the default value.  I will
use these setting for a bit and see if I like it.

type: sh