Examine file content to set ft=gitcommit
arc opens a tmp file, so we cannot use the file name.  This
is not a robust test, but good enough.  This examines the
content of the file, and sets the filetype to gitcommit
if no file type has yet been detected and if the final
line of the file is the string "-->"

type: git