Check mtime of HISTFILE
What a pain.  In bash 3.2.57(1) (macos), history -a does not
update HISTFILE. At all.  It does not fail....it just does
nothing.  So I am going to explicitly check mtime for
the file using stat.  (I suspect the stat call is not
portable and I will need to tweak it on other platforms;
will probably want to use perl.)  If the file is not
updated, fall back to history -w.