Implement a timed cache on warnings
It is getting annoying seeing the same warning multiple times.
This massively violates my policy of keeping things simple, but having
to use ancient versions of bash (I would have like to use an
associative array for the cache, but need this to run in 3.57)
is getting really, really annoying.

type: bash