Create function to collect mtime
As expected, stat is not standardized.  This is not an
attempt to be standard, just an attempt to work on both
of my laptops.  The stat -f %m version works on macos,
and the stat --format %Y works on Manjaro.  (Reminder to
self; burniniate Manjaro. Arch is a stupid distribution,
and my experiment using it has failed.  Or, it has been
successful and confirmed that arch is not worthwhile.)

type: rc