Re-exec modern bash, if available
This feels a little bit risky, and we might get into an infinite loop
of re-exec.  Perhaps we should put a recursion limiter in place,
but I think the logic here already prevents it.  The idea is that
the system is running ancient bash, but perhaps the local PATH
can find a newer bash.  If it can, exec it.  Otherwise emit the warning
and continue.

type: bash