Do not attempt to work with HEAD; just abort
Special casing is a bad idea, and this script is really intended
to be used without an argument.  Aborting when given "HEAD" is
the correct behavior, and validating the output of show-ref
seems better than the current behavior of failing with an error
when we cannot find the upstream.

type: git