Do not substitute large integers for a branch
I am finding this remapping occasionally annoying, and this seems like
a reasonable solution.  A common error is doing things like
"git config foo.bar 0" where 0 gets remapped to "main", but also
if a large number is passed as an argument, it almost certainly is
not intended to be replaced with a branch name.