Copy config template into newly created repos
Sigh.  Hooks need to be symbolic links so that I can update
the hooks in one place.  But config needs to be local to each
repo, so we do not want that ability.  Also, git-init appends
to the end of the config, and if it is a symbolic link it
modifies the content in dotfiles.git, so instead we just copy
the file as a template in newly created repos.

type: git