Allow trailers to be specified in t
This allows a user to do "t=type=git g ciam msg" to automatically
insert a trailer for "type" at the end of the commit message.  Note,
this probably does not belong in commit-msg. The trailers ought to
be added before the user starts editing the commit message rather than
after, but my usual workflow is to not edit the file at all by
to specify on the command line, and this is really just intended as
a simple convenience.

type: hook