Handle recursion and strip leading whitespace
The primary use case is for when there is some sequence
of commands I want to repeat, and rather than hitting
up-arrow n times repeatedly, we can now just do "b n".
But we cannot replace the "b n" in the history list, so we
need to increment the count to find the right command in
history.

type: sh