Stop terminating early if output not a tty
It is quite reasonable to redirect to a file or a pipe.  Looking
at the log, it seems I only did this to enable not worrying about
accidentally calling the function, but it looks like this is
less of an issue now that much functionality has been moved
into scripts.

type: git