Fix foldlevels for diff
Setting a foldlevel is a bad idea, since it
will be reset every time we enter the buffer,
and I definitely want fold level to remain constant
when changing buffers.  Re-mapping the folds is
better.  This still is not quite how I want things
to look, but is a good start.

type: git