Add fold level for ---
We were initially only trying to fold diffs from git
where there was a 4-line header in which the first line
was ^diff followed by a line with oids followed by ^---
and ^+++ lines.  It is also common to have those first
two lines omitted, and this seems to work well with both
cases.

type: git