Ignore file name for indentation
This is a bit hacky.  A file named "foo/python/bar/Dockerfile" should
not be considered a python file, but the extant code hits on
the "python" in the path.  Perhaps we should retain the
basename, but for now we will just discard the entire
path which should cause the checks to fall back to
config core.whitespace.  It seems best to be cautious and
fall back to uncertainty in the filetype check.

type: git