Only clone 3 commits of history
It is easy to convert to a full repo, and there does not seem to
be any reason not to do this.  I am grabbing a lot of repos
recently and this laptop is small.  Surprisingly, a git gc --aggressive
was able to reduce llvm-project/.git significantly after a clone. I
would expect the clone would have already done significant
compression, but apparently not.

type: git