Prevent syntax highlighter from going crazy on *
With the standard syntax highlighting, a single * on a line will start
a highlight group that persists until the next blank line.  This
change prevents that behavior.  This is copy-pasted from ChatGPT and
has not been even remotely tested other than opening a file once
and glancing at the behavior.

type: vim