Purge comment lines with leading whitespace
I do not think this is relevant, but want to be sure.  It seems
that interactive rebase with re-word does not invoke the
commit-msg script in some versions of git, but as far as I
can tell it is being invoked for me, so I do not understand
why I still see invalid messages in the log.

type: git