Enable space-before-tab for .py files
Space before tab should be an error.  Since git diff-index still
does not treat leading whitespace properly, we cannot throw away
all of the checks in the pre-commit hook.  Why on earth leading
space that is smaller than a tab-width is not treated as an error
is utterly baffling to me.

type: git