Use core.hooksPath instead of links in .git/hooks
This is the first step in fixing the hook architecture.  The name
".git-template-links" now does not make any sense, as there are
no longer links to the hooks/ directory, but that is an aesthetic
detail.

The idea here is that we will call our hooks from hooksPath
and then call the regular hooks instead of requiring local-hooks.
I still like the idea of local-hooks, but mostly I want to
be able to use the regular mechanism if someone is configuring
standard hooks in a repo.  But now the existing repos are
potentially going to recurse into themselves, so we need
to protect against that.

type: git