Make ~/.config a symbolic link into dotfiles
This simplifies the dotfiles setup substantially, and gives
us the ability to more easily add common config into
dotfiles.  I am a little worried that this makes
it difficult to to local configuration, but if that happens we
can always break the link on the box in question.  This
is (as always) a work in progress.

type: git