Stop trying to catch certain whitespace errors
The idea was to flag leading spaces that are less than a tabwidth when
indent-with-non-tab is set as an error.  For some ungodly reason,
git diff-index allows that, and there is not setting to make it treat it
as an error.  But if we use the hammer of "git diff --cached | grep -E",
we completely lose per-file attributes.

type: git