Left truncate the trailer in the log
It is cleaner if the columns remain aligned.  At the moment, I am
choosing ltrunc because I have a project in which the default trailer
is a URL with an issue number at the end, and it is convenient to
see the tail. Perhaps we can add options to control the truncation,
but that is for another day.

Type: git