Handle pager cleanly
Well, for sufficiently broad definition of "cleanly".
git l1 is throwing a syntax error without this change, so
clearly our manipulation of the pager is garbage.  This
cruft was all a mis-guided attempt to use --headers in less,
but we really need to get in here and clean up.

Type: git