Use diff-highlight explicitly
git var GIT_PAGER does not honor the setting of pager.diff config,
and even if it did, we cannot executed it like this.  If GIT_PAGER
is set to "foo | bar", then this attempt at execution just executes
foo with the arguments "|" and "bar".  We need to eval to get a pipe,
and this is just silly.

Type: git