Attempt to infer working directory
When not in a git repo, we should be able to do
things like "git log /path/to/git/repo".  This
attempts to expand the path via the completion
mechanism.  It does not work, but at least
it tells the user to use -C.  Perhaps we
should incorporate this into the git wrapper.

Probably not worth the effort.

Type: git